+++ internal/structs/user.go
@@ -2,7 +2,7 @@ package structs
 
 // User struct inside database
 type User struct {
-	ID       uint `gorm:"column:user_id;primary_key"`
+	ID       int64 `gorm:"column:user_id;primary_key"`
 	Username string
 	Password string
 	Email    string
@@ -18,9 +18,9 @@ type User struct {
 
 // Settings represent the user's account settings
 type Settings struct {
-	ID        uint `gorm:"column:setting_id;primary_key"`
-	UserID    uint `gorm:"primary_key"`
-	ShowLikes bool `gorm:"column:show_likes" json:"show_likes"`
-	ShowNSFW  bool `gorm:"column:show_nsfw" json:"show_nsfw"`
-	NSFWPage  bool `gorm:"column:nsfw_page" json:"nsfw_page"`
+	ID        uint  `gorm:"column:setting_id;primary_key"`
+	UserID    int64 `gorm:"primary_key"`
+	ShowLikes bool  `gorm:"column:show_likes" json:"show_likes"`
+	ShowNSFW  bool  `gorm:"column:show_nsfw" json:"show_nsfw"`
+	NSFWPage  bool  `gorm:"column:nsfw_page" json:"nsfw_page"`
 }
